Module reorganization and assemble tests (#47)
