--- src/init.jl
@@ -1,9 +1,0 @@
-using Reexport: @reexport
-
-@reexport using Dictionaries: dictionary
-
-# =========
-# Utils
-# =========
-include("./Utils.jl")
-@reexport using .Utils
